+++ src/components/App.jsx
@@ -1,4 +1,4 @@
-import styles from '../css/components/App.module.scss';
+import styles from './App.module.scss';
 import Header from './header/Header';
 import HeadLine from './main/HeadLine';
 import News from './main/News';
