+++ src/components/App.module.scss
@@ -1,4 +1,4 @@
-@import '../variable.scss';
+@import './variable.scss';
 
 #wrap {
     margin: 0 auto;
