+++ src/components/header/Dates.module.scss
@@ -1,4 +1,4 @@
-@import '../../variable.scss';
+@import '../variable.scss';
 
 .date {
     font-size: 18px;
