+++ src/components/header/Header.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Logo from './Logo';
 import Dates from './Dates';
-import styles from '../../css/components/header/Header.module.scss';
+import styles from './Header.module.scss';
 
 export default function Header() {
     return (
