+++ src/components/header/Logo.module.scss
@@ -1,4 +1,4 @@
-@import '../../variable.scss';
+@import '../variable.scss';
 
 .logo {
     font-size: 1.8rem;
