+++ src/components/variable.scss
