+++ src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/App.jsx';
-import './css/index.scss';
+import './components/index.scss';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
     <React.StrictMode>
